some refactoring, more to go
