+++ html/index.html
@@ -168,13 +168,18 @@
               </tr>
               {{ end }} {{ end }}
             </table>
-            {{ if .Country }}
+            {{ if .Country }} {{ if .Service == "geoip" }}
             <p>
               This information is provided from the GeoLite2 database created by
               MaxMind, available from
               <a href="https://www.maxmind.com">www.maxmind.com</a>
             </p>
-            {{ end }} {{ if .Latitude }}
+            {{ end}} {{ if .Service == "ipstack" }}
+            <p>
+              This information is provided from
+              <a href="https://www.ipstack.com">www.ipstack.com</a>
+            </p>
+            {{ end}} {{ end }} {{ if .Latitude }}
             <div class="pure-u-1 pure-u-md-1-1">
               <h2>Map</h2>
               <iframe
